Stop tracking Relaxed.Ruby.Style in repo
Relaxed.Ruby.Style is a good set of rules for most projects, but because
Rubocop changes the names and namespaces of rules over time, it's
generally inconvenient to have the style guide copied into a project
unless you lock yourself to a specific version of rubocop.

Rather than do that, it's easier to include the style guide as a gem so
that we're using an up to date version of everything that we know will
be compatible.

Additionally, this specifies a version of Rubocop that's after this CVE
was fixed, so GitHub will stop complaining about the vulnerability:
https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2017-8418